- Use [u]intptr_t casts to convert pointers to integers.

- Change IS_ERR* to return bool instead of long.

Reviewed by:	manu
Obtained from:	CheriBSD
Sponsored by:	DARPA
Differential Revision:	https://reviews.freebsd.org/D27577

+++ sys/compat/linuxkpi/common/include/linux/err.h
@@ -48,7 +48,7 @@ ERR_PTR(long error)
 static inline long
 PTR_ERR(const void *ptr)
 {
-	return (long)(intptr_t)ptr;
+	return (intptr_t)ptr;
 }
 
 static inline bool
